add vector float+ equilateral triangle validation

+++ line/vector.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 
 class Vector:
-    def __init__(self, x: int, y: int):
+    def __init__(self, x: float, y: float):
         self.x = x
         self.y = y
 
@@ -11,7 +11,7 @@ class Vector:
         y = self.y + other.y
         return Vector(x, y)
 
-    def __mul__(self, other: Vector) -> int:
+    def __mul__(self, other: Vector) -> float:
         return self.x * other.x + self.y * other.y
 
     def __eq__(self, other: object) -> bool:
@@ -20,7 +20,7 @@ class Vector:
         return self.x == other.x and self.y == other.y
 
     @property
-    def square_length(self) -> int:
+    def square_length(self) -> float:
         return self.x ** 2 + self.y ** 2
 
     @property
